use guint for port instead of guint16; fixes memory overrun and nasty
2007-05-31  Kristian Rietveld  <kris@gtk.org>

	* src/gossip-account-widget-jabber.c (account_widget_jabber_setup):
	use guint for port instead of guint16; fixes memory overrun and nasty
	crash.


svn path=/trunk/; revision=2443

+++ src/gossip-account-widget-jabber.c
@@ -707,7 +707,7 @@ account_widget_jabber_setup (GossipAccountWidgetJabber *settings)
 	GossipSession      *session;
 	GossipProtocol     *protocol;
 	GossipAccountParam *param;
-	guint16             port;
+	guint               port;
 	gchar              *port_str; 
 	const gchar        *id;
 	const gchar        *resource;
